Separate out most network partition support code
