+++ src/core/ddsi/include/dds/ddsi/q_misc.h
@@ -48,22 +48,12 @@ inline nn_sequence_number_t toSN (seqno_t n) {
 
 unsigned char normalize_data_datafrag_flags (const SubmessageHeader_t *smhdr);
 
-#ifdef DDS_HAS_NETWORK_PARTITIONS
-int WildcardOverlap(char * p1, char * p2);
-#endif
-
 extern const ddsi_guid_t nullguid;
 bool guid_prefix_zero (const ddsi_guid_prefix_t *a);
 int guid_prefix_eq (const ddsi_guid_prefix_t *a, const ddsi_guid_prefix_t *b);
 int guid_eq (const struct ddsi_guid *a, const struct ddsi_guid *b);
 int ddsi2_patmatch (const char *pat, const char *str);
 
-#ifdef DDS_HAS_NETWORK_PARTITIONS
-struct ddsi_config;
-struct ddsi_config_partitionmapping_listelem;
-struct ddsi_config_partitionmapping_listelem *find_partitionmapping (const struct ddsi_config *cfg, const char *partition, const char *topic);
-int is_ignored_partition (const struct ddsi_config *cfg, const char *partition, const char *topic);
-#endif
 #ifdef DDS_HAS_NETWORK_CHANNELS
 struct ddsi_config;
 struct ddsi_config_channel_listelem;
